+++ src/Session/OnMessageTask.cc
@@ -1,5 +1,5 @@
 /* This file is part of the CARTA Image Viewer: https://github.com/CARTAvis/carta-backend
-   Copyright 2018, 2019, 2020, 2021 Academia Sinica Institute of Astronomy and Astrophysics (ASIAA),
+   Copyright 2018-2022 Academia Sinica Institute of Astronomy and Astrophysics (ASIAA),
    Associated Universities, Inc. (AUI) and the Inter-University Institute for Data Intensive Astronomy (IDIA)
    SPDX-License-Identifier: GPL-3.0-or-later
 */
@@ -68,10 +68,5 @@ OnMessageTask* RegionDataStreamsTask::execute() {
 
 OnMessageTask* SpectralProfileTask::execute() {
     _session->SendSpectralProfileData(_file_id, _region_id);
-    return nullptr;
-}
-
-OnMessageTask* OnSplataloguePingTask::execute() {
-    _session->OnSplataloguePing(_request_id);
     return nullptr;
 }