+++ src/Session/OnMessageTask.h
@@ -1,5 +1,5 @@
 /* This file is part of the CARTA Image Viewer: https://github.com/CARTAvis/carta-backend
-   Copyright 2018, 2019, 2020, 2021 Academia Sinica Institute of Astronomy and Astrophysics (ASIAA),
+   Copyright 2018-2022 Academia Sinica Institute of Astronomy and Astrophysics (ASIAA),
    Associated Universities, Inc. (AUI) and the Inter-University Institute for Data Intensive Astronomy (IDIA)
    SPDX-License-Identifier: GPL-3.0-or-later
 */
@@ -105,15 +105,6 @@ class SpectralProfileTask : public OnMessageTask {
 public:
     SpectralProfileTask(Session* session, int file_id, int region_id) : OnMessageTask(session), _file_id(file_id), _region_id(region_id) {}
     ~SpectralProfileTask() = default;
-};
-
-class OnSplataloguePingTask : public OnMessageTask {
-    OnMessageTask* execute() override;
-    uint32_t _request_id;
-
-public:
-    OnSplataloguePingTask(Session* session, uint32_t request_id) : OnMessageTask(session), _request_id(request_id) {}
-    ~OnSplataloguePingTask() = default;
 };
 
 } // namespace carta