+++ test/BackendModel.h
@@ -1,5 +1,5 @@
 /* This file is part of the CARTA Image Viewer: https://github.com/CARTAvis/carta-backend
-   Copyright 2018, 2019, 2020, 2021 Academia Sinica Institute of Astronomy and Astrophysics (ASIAA),
+   Copyright 2018-2022 Academia Sinica Institute of Astronomy and Astrophysics (ASIAA),
    Associated Universities, Inc. (AUI) and the Inter-University Institute for Data Intensive Astronomy (IDIA)
    SPDX-License-Identifier: GPL-3.0-or-later
 */
@@ -54,8 +54,6 @@ public:
     void Receive(CARTA::CatalogFilterRequest message);
     void Receive(CARTA::StopMomentCalc message);
     void Receive(CARTA::SaveFile message);
-    void Receive(CARTA::SplataloguePing message);
-    void Receive(CARTA::SpectralLineRequest message);
     void Receive(CARTA::ConcatStokesFiles message);
     void Receive(CARTA::StopFileList message);
     void Receive(CARTA::SetSpatialRequirements message);